Add Acceptance Test for Deposit and Withdraw functions.

+++ cd-parent/cd-service/src/main/java/com/vince/cd/service/SystemService.java
@@ -5,7 +5,7 @@ import com.vince.cd.model.User;
 public interface SystemService {
 	User login(User user) throws Exception;
 	double getBalance(int accountNumber) throws Exception;
-	double deposit(int accountNumber, int amount) throws Exception;
-	double withdraw(int accountNumber, int amount) throws Exception;
+	double deposit(int accountNumber, double amount) throws Exception;
+	double withdraw(int accountNumber, double amount) throws Exception;
 	User getUserDetail(String username) throws Exception;
 }
